Latex issue fixed
