+++ front/package.json
@@ -15,22 +15,17 @@
     "@emotion/styled": "^11.11.5",
     "@reduxjs/toolkit": "^2.2.1",
     "@tailwindcss/postcss": "^4.0.12",
-    "better-react-mathjax": "^2.1.0",
-    "caniuse-lite": "^1.0.30001679",
     "formik": "^2.4.5",
-    "highlight.js": "^11.11.1",
     "i18next": "^23.10.1",
     "i18next-browser-languagedetector": "^7.2.0",
     "jspdf": "^2.5.1",
     "katex": "^0.16.21",
     "lucide-react": "^0.474.0",
-    "mathjax-full": "^3.2.2",
     "mermaid": "^11.4.1",
     "papaparse": "^5.5.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^14.1.0",
-    "react-latex-next": "^3.0.0",
     "react-loading-skeleton": "^3.4.0",
     "react-markdown": "^8.0.7",
     "react-redux": "^9.1.0",
@@ -42,13 +37,11 @@
     "redux": "^5.0.1",
     "redux-persist": "^6.0.0",
     "regenerator-runtime": "^0.14.1",
-    "rehype-highlight": "^7.0.2",
     "rehype-katex": "^6.0.3",
     "rehype-raw": "^6.1.1",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
-    "uuid": "^11.0.2",
-    "yup": "^1.4.0"
+    "uuid": "^11.0.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.56",
